Migrate backoffice to app router
