+++ app/backoffice/bookmarks/page.tsx
@@ -1,25 +1,12 @@
-import { GetServerSidePropsContext, NextPage } from 'next'
 import {
   EntriesList,
   GeneralizedEntry,
 } from '../../../components/backoffice/entries-list'
 import { prisma } from '../../../lib/utill/db'
-import { bookmark } from '@prisma/client'
 import { EntriesListHeader } from '../../../components/backoffice/entries-list-header'
 
-export async function getServerSideProps(context: GetServerSidePropsContext) {
-  const bookmarks = await prisma.bookmark.findMany({
-    orderBy: { created_at: 'desc' },
-  })
-
-  return {
-    props: {
-      bookmarks: JSON.parse(JSON.stringify(bookmarks)),
-    },
-  }
-}
-
-const BookmarksBackofficePage = ({ bookmarks }: { bookmarks: bookmark[] }) => {
+export default async function BookmarksBackofficePage() {
+  const bookmarks = await getBookmarks()
   return (
     <>
       <EntriesListHeader
@@ -34,4 +21,10 @@ const BookmarksBackofficePage = ({ bookmarks }: { bookmarks: bookmark[] }) => {
   )
 }
 
-export default BookmarksBackofficePage
+const getBookmarks = async () => {
+  const bookmarks = await prisma.bookmark.findMany({
+    orderBy: { created_at: 'desc' },
+  })
+
+  return JSON.parse(JSON.stringify(bookmarks))
+}
